requirements.txt
Flask==2.2.2
tensorflow-cpu==2.13.0
numpy==1.23.3
Pillow==9.2.0
werkzeug==2.2.2
gunicorn==20.1.0
tensorflow-hub==0.13.0